# swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: "astronauts"
  description: Descrizione dell'api
  termsOfService: http://swagger.io/terms/
  contact:
    name: Swagger API Team
    email: foo@example.com
    url: http://madskristensen.net
  license:
    name: MIT
    url: http://github.com/gruntjs/grunt/blob/master/LICENSE-MIT
host: astronauti.herokuapp.com
basePath: /api
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /astronauts:
    get:
      description: |
        Get all astronauts.
      operationId: getallastronauts
      responses:
        200:
          description: ritorna tutti gli elementi astronauts
          schema:
            type: array
            items:
              $ref: '#/definitions/astronauts'
        404:
          description: not found
          schema:
            $ref: '#/definitions/Error'
    post:
      description: Crea un nuovo astronauts.  Duplicates are allowed
      operationId: addastronauts
      parameters:
        - name: astronauts
          in: body
          description: astronauts da aggiungere all'array
          required: true
          schema:
            $ref: '#/definitions/Newastronauts'
      responses:
        200:
          description: astronauts response
          schema:
            $ref: '#/definitions/astronauts'
        404:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /astronauts/{id}:
    get:
      description: ritorna astronauts con quel ID
      operationId: getastronautsID
      parameters:
        - name: id
          in: path
          description: ID of astronauts to fetch
          required: true
          type: string
      responses:
        200:
          description: astronauts response
          schema:
            $ref: '#/definitions/astronauts'
        404:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      description: deletes a single astronauts based on the ID supplied
      operationId: deleteastronauts
      parameters:
        - name: id
          in: path
          description: ID of astronauts to delete
          required: true
          type: string
      responses:
        200:
          description: astronauts deleted
        404:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      description: put a single astronauts based on the ID supplied
      operationId: putastronauts
      parameters:
        - name: id
          in: path
          description: ID of astronauts to modify
          required: true
          type: string
      responses:
        200:
          description: astronauts modified
        404:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
definitions:
  nomeModulo:
    allOf:
      - $ref: '#/definitions/Newastronauts'
      - properties:
          id:
            type: string
  Newastronauts:
    required:
      - firstName
      - lastName
      - isInSpace
    properties:
      firstName:
        type: string
      lastName:
        type: string
      isInSpace:
        type: boolean
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string